split some function out
